+++ crates/hale-codegen/src/stdlib/mod.rs
@@ -14,4 +14,5 @@ pub(crate) mod decimal;
 pub(crate) mod env;
 pub(crate) mod math;
 pub(crate) mod process;
+pub(crate) mod rand;
 pub(crate) mod time;
